refactor: update imports to use consolidated types file for Avatar, Button, IconButton, Input, LinearProgress, Loader, Skeleton, and Typography components
