+++ src/components/atoms/Avatar/index.ts
@@ -1,2 +1,2 @@
 export * from './Avatar';
-export * from './Avatar.types';
+export * from './types';
