+++ src/components/atoms/Avatar/types.ts
