+++ src/components/atoms/IconButton/index.ts
@@ -1,2 +1,2 @@
 export * from './IconButton';
-export * from './IconButton.types';
+export * from './types';
