+++ src/components/atoms/Input/index.ts
@@ -1,3 +1,3 @@
 export * from './Input';
-export * from './Input.types';
+export * from './types';
 export * from './ControlledInput';
