+++ src/components/atoms/LinearProgress/index.ts
@@ -1,2 +1,2 @@
 export * from './LinearProgress';
-export * from './LinearProgress.types';
+export * from './types';
