+++ src/components/atoms/Loader/index.ts
@@ -1,2 +1,2 @@
 export * from './Loader';
-export * from './Loader.types';
+export * from './types';
