+++ src/components/atoms/Loader/types.ts
