+++ src/components/atoms/Skeleton/types.ts
